requirements.txt
Flask==3.0.0
langchain==0.0.1
python-dotenv==0.19.1
openai==0.27.1